Use comment_form_default_fields filter
Use a filter on comment_form_default_fields rather than appending the
reCAPTCHA widget using the comment_form action.  This removes the need
for hacking the tabindex and moving the submit button around.

On the downside, it requires the theme to use the WordPress 3
comment_form tag rather than constructing the form manually.